fix: Remaining CI regressions on Postgres — port type, admin destroy, hit_ratio.
- monitors.port: unsignedSmallInteger still maps to signed smallint in
  Postgres (unsigned is emulated via a check constraint, the type range
  stays -32768..32767), so 65000 still threw SQLSTATE[22003]. Switch the
  migration to integer which accepts the full 0..65535 TCP range.

- Admin\UserController::destroy now blocks deletion of a user whose role
  level is equal or higher than the actor's. The existing test asserted
  that behaviour but no code enforced it — a real privilege-escalation
  bug that only SQLite's looser Feature-test harness concealed.

- WarmSiteController stats24h.hit_ratio: cast the round() result to
  float so the Inertia payload type is consistent across SQLite (float)
  and Postgres (int from integer division), unblocking the assertion
  that expected 70.0 identically.

+++ app/Http/Controllers/WarmSiteController.php
@@ -172,7 +172,7 @@ class WarmSiteController extends Controller
                 'runs_total' => $totalRuns24h,
                 'total_urls' => (int) ($stats24h->total_urls ?? 0),
                 'hit_ratio' => ($stats24h->total_urls ?? 0) > 0
-                    ? round(($stats24h->total_hits / $stats24h->total_urls) * 100, 1)
+                    ? (float) round(($stats24h->total_hits / $stats24h->total_urls) * 100, 1)
                     : null,
                 'avg_response_ms' => (int) ($stats24h->avg_response_ms ?? 0),
                 'total_errors' => (int) ($stats24h->total_errors ?? 0),
